Supporting code has been extracted out of test file.
